feat(design-system/TUX-1172): Renaming Primitive dist
