+++ packages/design-system/src/components/Skeleton/variations/SkeletonInput.module.scss
@@ -1,5 +1,5 @@
 @use '~@talend/design-tokens/lib/tokens';
-@use '../../Form/FormPrimitives/mixins' as Form;
+@use 'src/components/Form/Primitives/mixins' as Form;
 
 .skeleton-input {
 	height: Form.$standardInputHeight;
